refactor role related code and add some log
