+++ lib/deploy/deploy-by-tpl.js
@@ -9,13 +9,18 @@ const getTpl = require('../tpl');
 const validate = require('../validate/validate');
 const debug = require('debug')('fun:deploy');
 
+const { green } = require('colors');
+
 const readFile = util.promisify(fs.readFile);
 
 let {
   makeApi, makeApiTrigger, makeFunction,
-  makeGroup, makeOtsTable, makeOtsInstance, makeOtsTrigger,
-  makeRole, attachPolicyToRole, makeService, makeTrigger, makeSlsProject, makeLogstore, makePolicy
+  makeGroup, makeOtsTable, makeOtsInstance, makeOtsTrigger, makeService, makeTrigger, makeSlsProject, makeLogstore
 } = require('./deploy-support');
+
+let {
+  makeRole, attachPolicyToRole, makePolicy
+} = require('../ram');
 
 function extractFcRole(role) {
   const [, , , , path] = role.split(':');
@@ -182,12 +187,14 @@ async function deployLogstore(projectName, logstoreDefinition) {
 async function deployLogs(resourcesDefinition) {
   for (const [projectName, v] of Object.entries(resourcesDefinition)) {
     if ((v || {}).Type === 'Aliyun::Serverless::Log') {
+      console.log(`Waiting for log ${projectName} to be deployed...`);
       const properties = (v || {}).Properties;
 
       const description = properties.Description || '';
       await makeSlsProject(projectName, description);
       
       await deployLogstore(projectName, v);
+      console.log(green(`log ${projectName} deploy success` ));
     }
   }
 }
@@ -362,15 +369,21 @@ async function deploy(tplPath) {
 
   for (const [name, resource] of Object.entries(tpl.Resources)) {
     if (resource.Type === 'Aliyun::Serverless::Service') {
+      console.log(`Waiting for service ${name} to be deployed...`);
       await deployFcService(name, resource);
+      console.log(green(`service ${name} deploy success` ));
     } else if (resource.Type === 'Aliyun::Serverless::Api') {
+      console.log(`Waiting for api gateway ${name} to be deployed...`);
       await deployApigateway(name, {
         apiDefinition: resource,
         template: tpl,
         tplPath
       });
+      console.log(green(`api gateway ${name} deploy success` ));
     } else if (resource.Type === 'Aliyun::Serverless::TableStore') {
+      console.log(`Waiting for table store ${name} to be deployed...`);
       await deployTablestore(name, resource);
+      console.log(green(`table store ${name} deploy success` ));
     } else if (resource.Type === 'Aliyun::Serverless::Log') {
       // ignore, done by deployLogs
     } else {
