+++ lib/deploy/deploy-support.js
@@ -5,7 +5,6 @@ const util = require('util');
 
 const ALY = require('aliyun-sdk');
 
-const Ram = require('@alicloud/ram');
 const FC = require('@alicloud/fc');
 const ots = require('@alicloud/ots2');
 const Pop = require('@alicloud/pop-core');
@@ -14,6 +13,10 @@ const CloudAPI = require('@alicloud/cloudapi');
 const getProfile = require('../profile').getProfile;
 const zip = require('../zip');
 const debug = require('debug')('fun:deploy');
+
+let {
+  makeRole, attachPolicyToRole, makePolicy
+} = require('../ram');
 
 const readFile = util.promisify(fs.readFile);
 
@@ -65,15 +68,6 @@ const getCloudApiClient = async () => {
   });
 };
 
-const getRamClient = async () => {
-  const profile = await getProfile();
-
-  return new Ram({
-    accessKeyId: profile.accessKeyId,
-    accessKeySecret: profile.accessKeySecret,
-    endpoint: 'https://ram.aliyuncs.com'
-  });
-};
 
 const getSlsClient = async () => {
   const profile = await getProfile();
@@ -338,7 +332,7 @@ async function makeFunction({
 const triggerTypeMapping = {
   'Datahub': 'datahub',
   'Timer': 'timer',
-  'Http': 'http',
+  'HTTP': 'http',
   'Log': 'log',
   'OSS': 'oss'
 };
@@ -360,7 +354,7 @@ function getTriggerConfig(triggerType, triggerProperties) {
       cronExpression: triggerProperties.CronExpression,
       enable: triggerProperties.Properties.Enable
     };
-  } else if (triggerType === 'Http') {
+  } else if (triggerType === 'HTTP') {
     return {
       authType: (triggerProperties.AuthType).toLowerCase(),
       methods: triggerProperties.Methods
@@ -492,130 +486,6 @@ async function makeTrigger({
   }
 
   return trigger;
-}
-
-async function getRamRole(ramClient, roleName) {
-  try {
-    return await ramClient.getRole({
-      RoleName: roleName
-    }, { timeout: 10000 });
-  } catch (ex) {
-    if (ex.name !== 'EntityNotExist.RoleError') {
-      throw ex;
-    }
-  }
-}
-
-async function makePolicy(policyName, policyDocument) {
-  const ram = await getRamClient();
-
-  let exists = true;
-
-  try {
-    await ram.getPolicy({
-      PolicyType: 'Custom',
-      PolicyName: policyName
-    });
-  } catch (ex) {
-    if (ex.code !== 'EntityNotExist.Policy') {
-      throw ex;
-    } else {exists = false;}
-  }
-  
-  if (!exists) {
-    await ram.createPolicy({
-      PolicyName: policyName,
-      Description: 'generated by fc fun',
-      PolicyDocument: JSON.stringify(policyDocument)
-    });
-  } else {
-    // avoid limitExceeded.Policy.Version
-    await deletePolicyNotDefaultVersion(ram, policyName);
-
-    await ram.createPolicyVersion({
-      PolicyName: policyName,
-      PolicyDocument:  JSON.stringify(policyDocument), 
-      SetAsDefault: true
-    });
-  }
-}
-
-async function deletePolicyNotDefaultVersion(ram, policyName) {
-  const listResponse = await ram.listPolicyVersions({
-    PolicyType: 'Custom',
-    PolicyName: policyName
-  });
-
-  const versions = (listResponse.PolicyVersions || {}).PolicyVersion;
-  if (versions) {
-    for (let version of versions) {
-      if (version.IsDefaultVersion === false) {
-        await ram.deletePolicyVersion({
-          PolicyName: policyName,
-          VersionId: version.VersionId
-        });
-      }
-    }
-  }
-}
-
-async function attachPolicyToRole(policyName, roleName, policyType = 'System') { 
-  const ram = await getRamClient();
-  
-  const policies = await ram.listPoliciesForRole({
-    RoleName: roleName
-  });
-
-  var policy = policies.Policies.Policy.find((item) => {
-    return item.PolicyName === policyName;
-  });
-
-  if (!policy) {
-    await ram.attachPolicyToRole({
-      PolicyType: policyType,
-      PolicyName: policyName,
-      RoleName: roleName
-    });
-  }
-}
-
-async function makeRole(roleName, createRoleIfNotExist, description='FunctionComput Default Role', assumeRolePolicy) {
-
-  const profile = await getProfile();
-
-  const ram = await getRamClient();
-
-  var role = await getRamRole(ram, roleName);
-
-  if (!assumeRolePolicy) {
-    assumeRolePolicy = {
-      'Statement': [
-        {
-          'Action': 'sts:AssumeRole',
-          'Effect': 'Allow',
-          'Principal': {
-            'Service': [
-              'fc.aliyuncs.com'
-            ]
-          }
-        }
-      ],
-      'Version': '1'
-    };
-  }
-
-  if (!role && createRoleIfNotExist) {
-    role = await ram.createRole({
-      RoleName: roleName,
-      Description: description,
-      AssumeRolePolicyDocument: JSON.stringify(assumeRolePolicy)
-    });
-  } else if (!role) {
-    console.error(`role ${roleName} not exist`);
-    process.exit(-1);
-  }
-
-  return role;
 }
 
 async function makeGroup(group) {
@@ -895,5 +765,5 @@ async function makeOtsTable({
 module.exports = {
   makeApi, makeApiTrigger, makeFunction,
   makeGroup, makeOtsTable, makeOtsInstance, makeOtsTrigger,
-  makeRole, attachPolicyToRole, makeService, makeTrigger, makeSlsProject, makeLogstore, makePolicy
+  makeService, makeTrigger, makeSlsProject, makeLogstore
 };
